+++ portal/tests/test_teacher.py
@@ -20,6 +20,7 @@ from django.test import Client, TestCase
 from django.urls import reverse
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.wait import WebDriverWait
+from selenium.webdriver.support import expected_conditions as EC
 
 from portal.forms.error_messages import INVALID_LOGIN_MESSAGE
 from portal.tests.test_invite_teacher import FADE_TIME
@@ -32,6 +33,8 @@ from .utils.messages import (
     is_email_updated_message_showing,
     is_password_updated_message_showing,
 )
+
+from portal.tests.test_invite_teacher import WAIT_TIME
 
 
 class TestTeacher(TestCase):
@@ -656,10 +659,14 @@ class TestTeacherFrontend(BaseTest):
             self.go_to_homepage().go_to_teacher_login_page().login(joining_email, joining_password).open_classes_tab()
         )
 
-        class_button = page.browser.find_element_by_id("class_button")
+        class_button = WebDriverWait(self.selenium, WAIT_TIME).until(
+            EC.element_to_be_clickable((By.ID, "class_button"))
+        )
         class_button.click()
 
-        edit_student_button = page.browser.find_element_by_id("edit_student_button")
+        edit_student_button = WebDriverWait(self.selenium, WAIT_TIME).until(
+            EC.element_to_be_clickable((By.ID, "edit_student_button"))
+        )
         edit_student_button.click()
 
         title = page.browser.find_element_by_id("student_details")
@@ -681,18 +688,24 @@ class TestTeacherFrontend(BaseTest):
         invite_button.click()
 
         # Once invite sent test the make admin button
-        make_admin_button = page.browser.find_element_by_id("make_admin_button_invite")
+        make_admin_button = WebDriverWait(self.selenium, WAIT_TIME).until(
+            EC.element_to_be_clickable((By.ID, "make_admin_button_invite"))
+        )
         make_admin_button.click()
 
         assert page.element_exists((By.CLASS_NAME, "popup-box__msg"))
         time.sleep(FADE_TIME)
-        cancel_button = page.browser.find_element_by_id("cancel_admin_popup_button")
+        cancel_button = WebDriverWait(self.selenium, WAIT_TIME).until(
+            EC.element_to_be_clickable((By.ID, "cancel_admin_popup_button"))
+        )
         cancel_button.click()
 
         # Delete the invite and check if the form invite with
         # admin checked also makes a popup
         time.sleep(1)  # let the page scroll down
-        delete_invite_button = page.browser.find_element_by_id("delete-invite")
+        delete_invite_button = WebDriverWait(self.selenium, WAIT_TIME).until(
+            EC.element_to_be_clickable((By.ID, "delete-invite"))
+        )
         delete_invite_button.click()
 
         for key in invite_data.keys():
@@ -701,11 +714,15 @@ class TestTeacherFrontend(BaseTest):
         checkbox = page.browser.find_element_by_name("make_admin_ticked")
         checkbox.click()
 
-        invite_button = page.browser.find_element_by_id("invite_teacher_button")
+        invite_button = WebDriverWait(self.selenium, WAIT_TIME).until(
+            EC.element_to_be_clickable((By.ID, "invite_teacher_button"))
+        )
         invite_button.click()
         assert page.element_exists((By.CLASS_NAME, "popup-box__msg"))
         time.sleep(FADE_TIME)
-        cancel_button = page.browser.find_element_by_id("cancel_admin_popup_button")
+        cancel_button = WebDriverWait(self.selenium, WAIT_TIME).until(
+            EC.element_to_be_clickable((By.ID, "cancel_admin_popup_button"))
+        )
         cancel_button.click()
 
         # Non admin teacher joined - make admin should also make a popup
@@ -716,7 +733,9 @@ class TestTeacherFrontend(BaseTest):
         page.browser.execute_script("location.reload()")
         page.browser.execute_script("window.scroll(0,900)")
         time.sleep(FADE_TIME)
-        make_admin_button = page.browser.find_element_by_id("make_admin_button")
+        make_admin_button = WebDriverWait(self.selenium, WAIT_TIME).until(
+            EC.element_to_be_clickable((By.ID, "make_admin_button"))
+        )
         make_admin_button.click()
 
         assert page.element_exists((By.CLASS_NAME, "popup-box__msg"))